requirements.txt
statsmodels
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm
torch
torch_geometric
setuptools
certifi
boto3
dotenv
scipy
PyYAML
dill
optuna
mlflow
dagshub
dvc
fastapi
uvicorn
jinja2
pydantic
python-multipart
sqlalchemy
psycopg2-binary
xgboost
#-e .
